Table: split TableNG hooks tests alongside their modules
Breaks the 1711-line hooks.test.ts into four files matching the modules from the
previous commit. render.test.tsx already sits beside render.tsx and is unchanged.

Unlike utils.test.ts, this suite shared state across its describe blocks: a
setupData() closure used by nearly every one, plus an emptyFilterResult constant.
Both move to hooks/testHelpers.ts. That filename is deliberate — the repo's
eslint config exempts testHelpers.ts from the rules it relaxes for test files, so
the fixture body moves across verbatim, type assertions and all.

The describe blocks move unchanged, dedented one level now that the outer
`TableNG hooks` wrapper is gone. No file gains or loses a test: the Table folder
still runs 819 tests, now over 39 suites instead of 31.

Splitting also surfaces four hooks with no coverage at all, previously invisible
inside one large file: useColumnResize, useScrollbarWidth, useTypographyCtx and
useContentAwareWidths. Adding those tests is follow-up work, not part of a move.
